Rename ExchangeSyncCode to ProcessSyncCode
The screen and its collaborators handle any sync code type and drive
the sync protocol to completion, so "Exchange" was misleading (it also
collided with the exchange-code concept). Renames the activity, view
model, contract, layout, animation and their references.
