+++ sync/sync-impl/src/main/res/raw/sync_process_sync_code_animation.json
